Move UriPathCodec to contrib.pathy
